refactor(frontend): extract useTabPersistence composable (P2-2 F4)
Move the workspace-tab localStorage persistence cluster (snapshot computed,
debounced persist, restore + the persist watch) out of DataStudioNew.vue
into composables/useTabPersistence.js, injecting shared reactive state and
createTabState so refs stay identical. Composable is assembled after
createTabState (before onMounted), preserving the original runtime order.

Verified: focused composable unit test (persist/flush/restore/version-guard/
debounce, 5 cases) + DataStudioNew mount smoke green; npm run lint (0 errors),
npm run build. Behavior preserved (verbatim move; storageKey passes the same
isDemoMode-derived key).
